refactor(toggle & home): fixed toggle item to allow reuse, changed home page name
