+++ src/frontend/src/components/Toggle.vue
@@ -1,8 +1,8 @@
 <template>
   <Switch
-    v-model="powerpoint"
+    v-model="toggleValue"
     :class="[
-      powerpoint ? 'bg-rrblue-400' : 'bg-rrgrey-600',
+      toggleValue ? 'bg-rrblue-400' : 'bg-rrgrey-600',
       'relative inline-flex flex-shrink-0 h-6 w-11 border-2 border-transparent rounded-full cursor-pointer transition-colors ease-in-out duration-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-rrblue-200',
     ]"
   >
@@ -10,27 +10,11 @@
     <span
       aria-hidden="true"
       :class="[
-        powerpoint ? 'translate-x-5' : 'translate-x-0',
+        toggleValue ? 'translate-x-5' : 'translate-x-0',
         'pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow transform ring-0 transition ease-in-out duration-200',
       ]"
     />
   </Switch>
-  <!-- <Switch
-    v-model="powerpoint"
-    :class="[
-      powerpoint ? 'bg-indigo-600' : 'bg-gray-200',
-      'relative inline-flex flex-shrink-0 h-6 w-11 border-2 border-transparent rounded-full cursor-pointer transition-colors ease-in-out duration-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500',
-    ]"
-  >
-    <span class="sr-only">Use setting</span>
-    <span
-      aria-hidden="true"
-      :class="[
-        powerpoint ? 'translate-x-5' : 'translate-x-0',
-        'pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow transform ring-0 transition ease-in-out duration-200',
-      ]"
-    />
-  </Switch> -->
 </template>
 
 <script>
@@ -39,7 +23,7 @@ import { Switch } from "@headlessui/vue";
 
 export default {
   props: {
-    powerpoint: Boolean,
+    toggleValue: Boolean,
   },
   components: {
     Switch,
