+++ src/frontend/src/router/index.js
@@ -3,18 +3,8 @@ import { createRouter, createWebHistory } from "vue-router";
 const routes = [
   {
     path: "/",
-    name: "Home2",
-    component: () => import("../views/Home2.vue"),
-  },
-  {
-    path: "/home",
     name: "Home",
     component: () => import("../views/Home.vue"),
-  },
-  {
-    path: "/",
-    name: "Home2",
-    component: () => import("../views/Home2.vue"),
   },
   {
     path: "/about",
